// apollo-bookstore-client-with-hooks/src/components/publishers/__generated__/GetPublishers.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: GetPublishers
// ====================================================

export interface GetPublishers_publishers {
    __typename: 'Publisher';
    id: string;
    name: string;
}

export interface GetPublishers {
    publishers: GetPublishers_publishers[];
}